:recycle: Refactor: 음성녹음 코드 리팩토링
